+++ web/documentserver-example/csharp/Users.cs
@@ -58,7 +58,9 @@ namespace OnlineEditorsExample
             "Can’t print the file",
             "Can create new files from the editor",
             "Can see the information about Group2 users",
-            "Can’t submit forms"
+            "Can’t submit forms",
+            "Can't close history",
+            "Can't restore the file version"
         };
 
         static List<string> descr_user_0 = new List<string>()