+++ web/documentserver-example/nodejs/helpers/users.js
@@ -82,6 +82,8 @@ const descrUser3 = [
   'Can create new files from the editor',
   'Can see the information about Group2 users',
   'Can’t submit forms',
+  'Can’t close history',
+  'Can’t restore the file version',
 ];
 
 const descrUser0 = [