+++ web/documentserver-example/nodejs/public/javascripts/jscript.js
@@ -19,6 +19,15 @@
 var language;
 var userid;
 var directUrl;
+var Formats;
+
+window.onload = function () {
+    fetch('formats')
+        .then((response) => response.json())
+        .then((data) => {
+            Formats = data;
+        })
+}
 
 if (typeof jQuery != "undefined") {
     jq = jQuery.noConflict();
@@ -272,6 +281,38 @@ if (typeof jQuery != "undefined") {
         jq("#uploadSteps").after('<iframe id="embeddedView" src="' + url + '" height="345px" width="432px" frameborder="0" scrolling="no" allowtransparency></iframe>');
     });
 
+    jq(document).on("click", "#beginEditConverted:not(.disable)", function () {
+        var fileId = encodeURIComponent(jq('#hiddenFileName').attr("data"));
+        if (UrlEditor == "wopi-action"){
+            var url = UrlEditor + "/" + fileId + "?action=edit" + collectParams(true);
+        }else{
+            var url = UrlEditor + "?fileName=" + fileId + collectParams(true);
+        }
+        window.open(url, "_blank");
+        jq('#hiddenFileName').val("");
+        jq.unblockUI();
+        window.location = collectParams();
+    });
+
+    jq(document).on("click", "#beginViewConverted:not(.disable)", function () {
+        var fileId = encodeURIComponent(jq('#hiddenFileName').attr("data"));
+        if (UrlEditor == "wopi-action"){
+            var url = UrlEditor + "/" + fileId + "?action=view" + collectParams(true);
+        }else{
+            var url = UrlEditor + "?mode=view&fileName=" + fileId + collectParams(true);
+        }
+        window.open(url, "_blank");
+        jq('#hiddenFileName').val("");
+        jq.unblockUI();
+        window.location = collectParams();
+    });
+
+    jq(document).on("click", "#downloadConverted:not(.disable)", function () {
+        var fileId = jq('#hiddenFileName').attr("data");
+        if (jq("#downloadConverted").attr("data") == "fromConverter") window.location.assign(fileId);
+        else window.location.href = "download?fileName=" + encodeURIComponent(fileId);
+    });
+
     jq(document).on("click", ".reload-page", function () {
         setTimeout(function () { window.location = collectParams(); }, 1000);
         return true;
@@ -292,6 +333,97 @@ if (typeof jQuery != "undefined") {
             window.location = collectParams();
         }
     });
+
+    jq(document).on("click", ".convert-file", function () {
+        const currentElement = jq(this);
+        var fileName = currentElement.attr("data");
+        var type = currentElement.attr("data-type");
+
+        jq.blockUI({
+            theme: true,
+            title: "Converting file" + "<div class=\"dialog-close\"></div>",
+            message: jq("#convertingProgress"),
+            overlayCSS: { "background-color": "#aaa" },
+            themedCSS: { width: "539px", top: "20%", left: "50%", marginLeft: "-269px" }
+        });
+
+        jq("#convertFileName").text(decodeURIComponent(fileName));
+        jq("#convertFileName").removeClass("word slide cell");
+        jq("#convertFileName").addClass(type);
+        jq("#convTypes").empty();
+        let convExtensions = Formats.find(format => {return format.name == fileName.split('.').pop()}).convert;
+        convExtensions.forEach(ext => {
+            jq("#convTypes").append(jq(`<td name="convertingTypeButton" id="wordTo${ext}" class="button hoar" data="${ext}">${ext}</td>`));
+        });
+        jq("#hiddenFileName").val(fileName);
+        jq("#convertStep1").addClass("done");
+        jq("#convertStep2").addClass("waiting");
+    });
+
+    jq(document).on("click", "td[name='convertingTypeButton']:not(.disable, .orange)", function () {
+        const currentElement = jq(this);
+        let id = currentElement[0].id;
+        let fileExt = jq(`#${id}`).attr("data");
+        jq(`#${id}`).addClass("orange");
+        jq("td[name='convertingTypeButton']").addClass("disable");
+        jq("#convertStep2").removeClass("waiting").removeClass("done").addClass("current");
+        jq("#convertStep2").text('2. File conversion');
+        jq("#convert-descr").removeClass("disable");
+        jq("#convertPercent").text("0 %");
+        jq("#hiddenFileName").attr("placeholder",fileExt);
+        jq("#downloadConverted").addClass("disable");
+        jq("#beginEditConverted").addClass("disable");
+        jq("#beginViewConverted").addClass("disable");
+        mustReload = true;
+        
+        convertFile();
+    });
+
+    function convertFile (filePass) {
+        let fileName = decodeURIComponent(jq("#hiddenFileName").val());
+        let fileExt = jq("#hiddenFileName").attr("placeholder");
+
+        if (timer != null) {
+            clearTimeout(timer);
+        }
+        timer = setTimeout(function () {
+            jq.ajaxSetup({ cache: false });
+            jq.ajax({
+                async: true,
+                type: "post",
+                dataType: "json",
+                data: {filename: fileName, filePass: filePass, lang: language, fileExt: fileExt},
+                url: UrlConverter,
+                complete: function (data) {
+                    try {
+                        var response = jq.parseJSON(data.responseText);
+                    } catch (e)	{
+                        response = { error: e };
+                    }
+                    if(response.step !==100) {
+                        jq("#convertPercent").text(`${response.step} %`);
+                        convertFile();
+                    } else {
+                        jq("#convertPercent").text(`${response.step} %`);
+                        jq("#convertStep2").removeClass("current").addClass("done");
+                        jq("#convertStep2").text(`2. File conversion to ${fileExt}`);
+                        jq("#downloadConverted").removeClass("disable");
+                        if (response.error !== "FileTypeIsNotSupported") {
+                            jq("#hiddenFileName").attr("data",response.filename);
+                            jq("#beginEditConverted").removeClass("disable");
+                            jq("#beginViewConverted").removeClass("disable");
+                            jq("#downloadConverted").attr("data","fromStorage");
+                        } else {
+                            let newFilename = fileName.split('.').slice(0,-1).join('.')
+                            jq("#hiddenFileName").attr("data",response.filename.split("&filename=download").join(`&filename=${newFilename}`));
+                            jq("#downloadConverted").attr("data","fromConverter");
+                        }
+                        jq("td[name='convertingTypeButton']").removeClass("disable orange");
+                    }
+                }
+            });
+        }, 1000);
+    }
 
     jq(document).on("click", ".delete-file", function () {
         const currentElement = jq(this);