+++ web/documentserver-example/php/src/views/DocEditorView.php
@@ -247,46 +247,27 @@ final class DocEditorView extends View
         }
 
         $historyLayout = "";
+
+        if ($user->id == "uid-3") {
+            $historyLayout .= "config.events['onRequestHistoryClose'] = null;
+                config.events['onRequestRestore'] = null;";
+        }
+
         if ($user->id != "uid-0") {
             $historyLayout .= "// add mentions for not anonymous users
-                config.events['onRequestUsers'] = function (event) {
-                    if (event && event.data){
-                        var c = event.data.c;
-                    }
-                    switch (c) {
-                        case \"info\":
-                            users = [];
-                            var allUsers = {usersInfo};
-                            for (var i = 0; i < event.data.id.length; i++) {
-                                for (var j = 0; j < allUsers.length; j++) {
-                                    if (allUsers[j].id == event.data.id[i]) {
-                                        users.push(allUsers[j]);
-                                        break;
-                                    }
-                                }
-                            }
-                            break;
-                        case \"protect\":
-                            var users = {usersForProtect};
-                            break;
-                        default:
-                            users = {usersForMentions};
-                    }
-                    docEditor.setUsers({
-                        \"c\": c,
-                        \"users\": users,
-                    });
-                };
+                config.events['onRequestUsers'] = onRequestUsers;
                 // the user is mentioned in a comment
-                config.events['onRequestSendNotify'] = function (event) {
-                    event.data.actionLink = replaceActionLink(location.href, JSON.stringify(event.data.actionLink));
-                    var data = JSON.stringify(event.data);
-                    innerAlert(\"onRequestSendNotify: \" + data);
-                };
+                config.events['onRequestSendNotify'] = onRequestSendNotify;
                 // prevent file renaming for anonymous users
                 config.events['onRequestRename'] = onRequestRename;
                 // prevent switch the document from the viewing into the editing mode for anonymous users
-                config.events['onRequestEditRights'] = onRequestEditRights;";
+                config.events['onRequestEditRights'] = onRequestEditRights;
+                config.events['onRequestHistory'] = onRequestHistory;
+                config.events['onRequestHistoryData'] = onRequestHistoryData;";
+            if ($user->id != "uid-3") {
+                $historyLayout .= "config.events['onRequestHistoryClose'] = onRequestHistoryClose;
+                config.events['onRequestRestore'] = onRequestRestore;";
+            }
         }
         $this->tagsValues = [
             "docType" => getDocumentType($filename),